requirements.txt
streamlit
pandas
openai
google-generativeai
python-dotenv